requirements.txt
# Speech Diarization Benchmark Dependencies

# Core ML frameworks
torch>=2.0.0
torchaudio>=2.0.0

# Diarization models
pyannote.audio>=3.1.0
pyannote.metrics>=3.2.0
pyannote.core>=5.0.0
speechbrain>=1.0.0

# API clients
assemblyai>=0.23.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0


